requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.9.0
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
coqpit-config==0.2.1
coqui-tts==0.27.0
coqui-tts-trainer==0.3.1
cycler==0.12.1
cython==3.1.2
dateparser==1.1.8
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2025.7.0
greenlet==3.2.3
groq==0.30.0
grpcio==1.73.1
gruut==2.4.0
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
inflect==7.5.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.3.26
langchain-core==0.3.68
langchain-groq==0.3.6
langchain-openai==0.3.28
langchain-tavily==0.2.7
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.44.0
markdown==3.8.2
markupsafe==3.0.2
matplotlib==3.10.3
monotonic-alignment-search==0.2.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.0
openai==1.95.1
orjson==3.10.18
ormsgpack==1.10.0
packaging==24.2
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pyparsing==3.2.3
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
sqlalchemy==2.0.41
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.2
torch==2.7.1+cu128
torchaudio==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.53.2
typeguard==4.4.4
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
